requirements.txt
notebook~=7.0.2
transformers~=4.31.0
datasets~=2.14.4
ipython~=8.14.0
ipywidgets~=8.1.0
google-cloud-bigquery~=3.11.4
pandas~=2.0.3
python-dotenv~=1.0.0
torch~=2.0.1
google-cloud-bigquery-storage~=2.22.0
pyarrow~=12.0.1
db-dtypes~=1.1.1
accelerate~=0.21.0
scikit-learn~=1.3.0
nltk~=3.8.1
tensorboard~=2.14.0
python-lsp-server~=1.7.4
openai~=0.27.8
matplotlib~=3.8.0rc1
click~=8.1.7
pytest~=7.4.0
black~=23.7.0
responses~=0.23.3
